# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ottosfa/v2.2/merchant-new-recruitment/bulk": {
            "post": {
                "description": "Merchant New Recruitment - Upload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant New Rec v2.2"
                ],
                "summary": "Merchant New Recruitment - Upload",
                "operationId": "Merchant New Recruitment - Upload v2.2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with file csv",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant New Recruitment - Upload EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/activitas-salesmen/list": {
            "post": {
                "description": "Activitas Salesmen - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activitas Salesmen V2.3"
                ],
                "summary": "Activitas Salesmen - List",
                "operationId": "Activitas Salesmen - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.ListActivitasSalesmenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activitas Salesmen - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2.3/activitas-salesmen/lits-detail-callplan": {
            "post": {
                "description": "Activitas Salesmen - List Detail Callplan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activitas Salesmen V2.3"
                ],
                "summary": "Activitas Salesmen - List Detail Callplan",
                "operationId": "Activitas Salesmen - List Detail Callplan",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.DetailListActivitasTodolistReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activitas Salesmen - List Detail Callplan",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2.3/activitas-salesmen/lits-detail-todolist": {
            "post": {
                "description": "Activitas Salesmen - List Detail Todo List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activitas Salesmen V2.3"
                ],
                "summary": "Activitas Salesmen - List Detail Todo List",
                "operationId": "Activitas Salesmen - List Detail Todo List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.DetailListActivitasTodolistReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activitas Salesmen - List Detail Todo List",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/gender/": {
            "get": {
                "description": "Gender - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gender V2.3"
                ],
                "summary": "Gender - List",
                "operationId": "Gender - List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/jobcategories/save": {
            "post": {
                "description": "JobCategories - Create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JobCategories"
                ],
                "summary": "JobCategories - Create",
                "operationId": "JobCategories - Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.JobCategories"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JobCategoriess - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/todolist/merchant-list": {
            "post": {
                "description": "Merchant List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList V2.3"
                ],
                "summary": "Merchant List",
                "operationId": "merchant-list",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MerchantListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/merchant-new-recruitment/bulk": {
            "post": {
                "description": "Merchant New Recruitment - Upload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant New Rec"
                ],
                "summary": "Merchant New Recruitment - Upload",
                "operationId": "Merchant New Recruitment - Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with file csv",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant New Recruitment - Upload EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dbmodels.DetailListActivitasTodolistReq": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "integer",
                    "example": 2
                },
                "date": {
                    "type": "string",
                    "example": "2020-11-20"
                },
                "id": {
                    "type": "string",
                    "example": "9957"
                },
                "keyword": {
                    "type": "string",
                    "example": "OP1B00027785"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "Not Exist"
                }
            }
        },
...
        "dbmodels.JobCategories": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dbmodels.ListActivitasSalesmenReq": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string",
                    "example": "SO TIMUR Indomarco"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "period_from": {
                    "type": "string",
                    "example": "2021-02-05"
                },
                "period_to": {
                    "type": "string",
                    "example": "2021-02-10"
                },
                "sales_type_id": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "models.MerchantListReq": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                }
            }
        },
...
        "models.Meta": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                }
            }
        },
...
    }
}